Improve layout preview responsiveness
By default, the NlPreviewForm has used an 800ms MergingUpdateQueue to
avoid issuing too many repaints when writing XML. The same delay was
used for the model rendering when dragging and dropping components on
the right.
This CL creates a new queue to handle XML editing with 250ms delay and
leaves the 10ms delay when dragging and dropping components on the
right panel. The new MergingUpdateQueue has also restart enabled and
will start the 250ms count every time the caret moves (so we don't need
the 800ms delay).

Change-Id: Ie1199be78b686ae8e492018eaaa5ea0aaa27fa95