[New PSD] Created "Select Module" dialog.
This dialog is generic, client code can specify the title of the
dialog, and what to do when the user presses the 'OK' button.

In this CL we use this dialog to have the user select a module
when clicking the "Add Library Dependency" action in the
"All Modules" dependencies view.

Change-Id: I830309e1c0b5848e903289279b5b1ae0a08bd8c0