Show ranges of atoms in groups of atoms.
Currently if i want to find somewhere i was looking it takes
a long time to get to the correct id, and involves lots of opening
and closing the groups to see the numbers inside. With this change
the user can now see at a glance, what ids of draw call are inside
each group.

Change-Id: I83072ef98a2006c5492df3c51b1d27a44d1e9c83