[New PSD] "Target Modules" now only show the modules and the scope of a dependency.
Now the "Target Modules" view in the "All Modules" dependencies configurable
show the module affected by a dependency, but instead of showing all the
variants/artifacts that contain that dependency, the view shows the scope of
the dependency and whether the dependency is "declared" or "transitive".

This simplifies the UI greatly, given that in a project with more than 1
product flavor, the "target Modules" view can be too cluttered.

Change-Id: Ib20458f9d9053f240975638372237e3b5a8915ef