Create new layoutlib module
This creates a new layoutlib module that contains the layoutlib.jar
library dependency and a few classes required to load it dynamically.
Only designer and android modules will depend on this new layoutlib
module.

For now, I'll leave layoutlib.jar as NOT exported since we are not
making use of any of the classes just yet. Also removing a few
accidental usages of Lists from layoutlib.jar instead of the Guava
version.

Change-Id: Ie8e044eab144882534e18c300f01bde60d38032d