A round of UX tweaks to tutorial elements.
Make a label that acts as a toggle have a hand cursor to
indicate clickable.
Correct numeric indicator font face to be consistent with other
elements.
Add some padding to code blocks that got accidentally removed in
the switch to using markup display.
Changing the tutorial links (which are really buttons) to look
more like links.

Change-Id: Ic7f07013c3e65faabd5eff17263bb10aa1812507