[New PSD] Removed automatic selection of matching nodes in trees.
This CL removes automatic selection of matching nodes. For
example, selecting "junit:4.12" in the "Resolved Dependencies" view
will automatically trigger a search for other "junit:4.12" nodes
in the tree, and the view will select all the found nodes.

This can be confusing for the user, especially because trees scroll
up and down when doing the automatic selection.

The automatic selection is not automatic any more, but it can
be manually trigger using a new toolbar button action, added in
this CL as well.

Change-Id: I1318ef71cae52cc04f7a4fd0c998c560b1557772