View Layout actions toolbar
This CL adds a layout actions toolbar, below the configuration toolbar,
which provides actions relevant to the currently selected widgets,
their parent, or if there's no selection, the root layout.

It also registers sample view actions for the LinearLayout view handler.

This CL also prepares for the new toolbar layout, but doesn't update
the toolbar order yet (and doesn't add the new zoom actions etc.)
This will be added in a followup CL.

Change-Id: I03d4c671cc293e7611263ae1f430f359e9ee80c6