Improve ChooseDeviceDialogFixture
ChooseDeviceDialogFixture.find could return the dialog, while it
was still loading the devices list (indicated by the spinning icon
in the top right corner of the list). In this case, some actions
(in my case, clicking the OK button) do not produce the expected
result. I considered adding a "waitForDevicesToLoad" function
instead, but I decided that it was not worth the trouble as
probably noone will be interested in interacting with a dialog
while devices are loading.

Change-Id: I2c64c97a3573f30fea09e7be2a470338b6a7c937