Minimize overlap problems when resizing the property panel.
When the property panel is resized horizontally there is a problem
with space such that rows with multiple components are squeezed
on top of each other.
This CL does not eliminate this problem but helps by:
- Reconfiguring the split components onto separate lines when space
  gets tight.
- Allow combo boxes to resize smaller than the popup.
- Make the controls grow to the end of the properties panel.

Also: add a little space on either side of the edit area of the
NlReferenceEditor. It used to positioned too close to the
resource inspector button and the color icon.

Change-Id: I6a428c878cccaa7db9c63a76dd037da4e4b53b7a