Migrate towards a reusable InstantRunBuilder
The logic for identifying what gradle tasks to run was spread across the
run configuration and the build logic. This CL attempts to unify them
all into a single InstantRunBuilder class, which could potentially be
reused from different contexts.

This class receives as input the target device, the actions taken from
the run UI and the project context. Using these inputs, it identifies
the type of build that needs to happen (incremental, full or clean) and
issues the appropriate commands with the correct arguments to Gradle.

The incremental tasks might need a rebuild in case there are no
artifacts that were generated. This logic was also spread across
multiple classes earlier, but now it is easily handled by the builder.

The other major area of cleanup is that various methods in
InstantRunManager have been fixed to use the same context objects to
obtain information they need as opposed to assuming a Gradle based
project.

b/28170881

Change-Id: I61403717303679251e5cb98e2120821af15122d1