Removing enclosing scroll pane on code sample component.
Code is now displayed in an editor which has its own scroll pane.
Due to the nested scroll panes, click-drag and trackpad/scrollwheel
scrolling was being handled by the interior scroller and not
bubbling. As the interior scroll thought it had all the space it
needed, this was treated as a no-op.

Switched to overriding the scroll pane inside the editor in a
similar fashion to the way the enclosing scroller was set up.

Note that this includes some order-of-events hacks but this appears
necessary due to access issues to the inner scroll pane.

Change-Id: Ia5801b59872c90316f1a4d2eb9c842e24eedb952