Handle custom namespaces on component drop
This CL updates the component model such that when it's inserting XML
fragments into the document after they've been customized by view
handlers, it ensures that all the namespaces defined in the component
document are also imported in the layout document.

Similarly, if the namespace *is* imported, but under a different
prefix, it renames the prefixes in the target XML fragment before
importing.

Finally, if the namespace isn't imported but the desired prefix is
already used for a different namespace, a different available prefix
is chosen and again all the attribute references are updated before
document insertion.

This fixes the bug where you for example make a new layout from the
Project view's context menu (which doesn't pre-define the app
namespace), and then drag a buton into a ConstraintLayout. This would
place the button at (0,0) since the right constraint namespace was not
imported.

Change-Id: I50ad85f503d04a822be212934697c979f00e1c34