LayoutPreviewTest: testRenderingDynamicResources
This tests keeps failing on Jenkins, but the screenshot looks OK.
It shows a "Sync Now" link, but the tests fails with a "not displaying"
error...
During the typing, the error message is being shown multiple times.
The problem happens because we grab the Panel of one message,
that then is dismissed, and replaced by another that looks similar.

Change-Id: If577b0f548e1a91f89fa4e74171cb3163887907f