Add a resize event to fix layout issues.
Due to some sort of race condition calculating component sizes, there
are various clipping and overlapping issues in tutorial content. The
only known item that fixed the problem was resizing the pane. In the
short term, just trigger a resize event at the end of the card
constructor as a bandaid fix.

Change-Id: Ia2d9ca0f1b348dfae213d73fb56592ceca93f273