Check isActive instead of callback in ToolWindowFixture.activate
GradleTasksTest@testTaskCancellation gets stuck waiting for the
Tool Window to get active.
If the Tool Window has a pending request to get active,
ToolWindowFixture@activate() will try to activate it again
(isActive() is false) and the Call back is not called, as the request
is ignored.
This could be the same problem affecting other tests that
sometimes fail to open the tool window.

Change-Id: Ia52f369eb9d66af1af6be48853b436f4476b1db0