Fix layout editor tag identity handling
If you perform some structural edits (such as dragging a widget from
one layout to another), the IDE preserves PSI identity pretty well,
and the layout editor model sync code uses this as the strongest
signal to preserve component hierarchy.

However, if you *Undo* this operation, the PSI elements are reused in
a strange way (such that a PSI tag which used to point to a
<LinearLayout> is now suddenly its child <Button> instead.)  This
*really* confused the model sync code.

This fix handles this by checking tag name identity before reusing
components mapped to PSI elements, and refusing to recycle objects
where the tag names have changed.

(This also fixes the related (same cause) bug where you could switch
to the XML editor and edit the tag name from RelativeLayout to
LinearLayout and still have component state cached, such as using the
RelativeLayoutHandler instead of the LinearLayoutHandler during edit
operations.)

Change-Id: I3911503d0ed818a66dc6d6a1aab4c4f3dd7caeb6