Now we handle BaseConfig#getVersionNameSuffix safely.
This is a new method added in builder-model 2.2. We need
to safely invoke it (using a try/catch) so uses of older
versions of the plugin won't throw exceptions.

Even though we have a check for version, we have 2.2.0
with and without this method, internally, making it
unnecessarily painful to work.

Change-Id: I5217e43df76fcbaf2e4a9afd3ad3c142ad8f02cb