Ensure that the editor type is preserved on file match
If you have two separate versions of the same layout, such
as portrait and landscape versions, and you click on the
orientation toggle, we automatically switch to the best matching
file.

This CL ensures that we stay in the same editor type -- e.g.
if you're in XML editor mode, we switch to XML editor mode in
the other variation -- and ditto for the layout editor.

Change-Id: Ie4c83cb87b4a77e2b371cc07cfe4e1558d68c138