Add a context menu to the design surface and outline
This CL cleans up the action handling a bit, and adds a context menu
to the design surface (and to the structure pane) which consists of:

* actions contributed by the selected component (and this CL now
  ensures that if you right click over a point that is not part of the
  selection, the selection is replaced by the component under the
  mouse)

* actions contributed by the parent view group

* global actions (such as a new Selection submenu where you can select
  parent, siblings, same types, etc) and copy/cut/paste.

We'll add more items to the menu soon.

This CL also changes the addViewActions API on ViewHandlers to now be
split up into two separate methods: addToolbarActions() and
addPopupMenuActions().

Finally, it starts caching ViewActions (which the API already promised
it would do) so ViewActions really can not store any instance state
now.

Change-Id: I07bdbe4b8199589c54d369bb540af8b51cd83d77