[New PSD] Replaced "Resolved Dependencies" view with "Target Artifacts".
This CL replaces the "Resolved Dependencies" view with the new
"Target Artifacts" view in the "Module Dependencies" perspective.

The reason is that having there is a lot of information to display
in the "Resolved Dependencies" view for non-trivial projects. When
editing a dependency it makes more sense to show only information
related to the selected dependency, instead of information related
to the whole module, which can be overwhelming.

Instead, now we show the variants and artifacts that contain the
selected dependency. I'll move the "Resolved Dependencies" view
as a top-level perspective in the next CL.

Change-Id: Ic917b6e74cca6fd241d8bd804e4dfe51d4df28f7