Fix missing width/height cache
When a widget is loaded in wrap_content or spring mode we don't have
a suitable width or height attribute we can use as the fixed dimension.
This change simply uses the model's current width and height, converted
to dp, as a fallback value. It works great in practice as it makes the
system more stable when cycling through dimensions.

Change-Id: I92e7dfa5649884bf3c2094c54d17a8fad2f0f594