Show resource table when viewing arsc files
When an ARSC file is selected, we display the resource table stored in
that ARSC file. The UI shows a list on the left that shows the various
resource types (id, string, attr, ..), and when a particular resource
type is selected, a table of resources of the selected type and their
values in various configurations is displayed.

Change-Id: Id9f87be2c17dc02a644012290677ee659cd264a2