Remove suffixes heuristic and unnecessary array wrapping
Code completion used to look up styleables with names resulted in
appending a bunch of suffixes to simple class names, which were
_MarginLayout, _Layout and _Cell. Since _MarginLayout and _Cell
are occurring only once in Android framework's attrs.xml, this is
just wasteful computation.

This CL removes trying all these suffixes for each possible view,
which makes it possible to remove a bunch of methods that iterate
over arrays. This is done here as well and gets rid of a bunch of
excess allocations.

Change-Id: I95e4f088bbb820ba14d0edccc7129df3561717c1