[New PSD] Simplified "Resolved Dependencies" view.
With this CL, instead of showing a single massive
tree view that displays all variants, artifacts and
dependencies, the "Resolved Dependencies" view
splits this view in 2, similar to the "All Projects"
dependencies view. On the left we show all
dependencies, and on the right we show the
variants and artifacts where the selected dependency
is.

Many thanks to Thiru for this great suggestion.

Change-Id: I3fab71500d95546974243c07b187a90cd688cf1e