Handle non android launch scenarios in before run provider
A recent refactoring (91bfb783d8f5) made an assumption that the android
run config context will always be available with a set of devices while
a build happens.

This is incorrect in a few scenarios such as:
 - running unit tests
 - running tests on the cloud

This CL fixes NPEs that happened in such circumstances.

Fixes https://code.google.com/p/android/issues/detail?id=209295

Change-Id: Ie8d51ebe7cd6a0ab2c7669893e9910ff842fbc89