Support deletion in the component outline
Also force the update in the outline to happen more quickly
(rather than the usual 250ms delay to reflect model updates)
when the user invokes delete in the compute tree, since otherwise
it feels sluggish to press delete and see the component stay
around for another 250+ ms. This is also done for Drag & drop
operations in the outline.

Change-Id: If5f94673cfb9704172a04b47463cb75d0c05947d