Adding "Refresh External C++ Projects" action.
This action will enabled only when the native modules are present
in the project (the action will be enabled only when the external
C++ projects present in the project in a followup change).

The action will sync the project with an additional argument to
indiacate Gradle to regenerate the external build systems json file
before syncing the project.

Change-Id: Ifa05247c2fc071a2bf258b0f398f8ff9284447e0