Activate the text editor when double click on a design component.
If this is the preview design surface goto the correct spot in the
xml file.
Otherwise: look through the current properties and find the
preferred property and open the editor. This is normally the text
property.
In this CL this is implemented for the simple property inspector
but not for the advanced property sheet. This will be done in a
followup CL. This should be good enough for IO...

Change-Id: Ia34ac59ceebfd7aba759dda80640e69f218f72c1