Add context filtering.
Use the new "contexts-hierachies" feature to allow the user to select
a context to filter the command list and filmstrip.

The client currently assumes each context only has a single hierarchy,
which is currently true. Once GAPIS produces multiple hierarchies per
context, we should offer a hierarchy selection as well.

This CL should be backwards compatible with earlier versions of GAPIS.

Change-Id: I67d9c217fe742aefd180cd7901d0bfd42be94bd3