Prevent broken rendering on startup due to missing AndroidModel
If you restart the IDE the layout editor *sometimes* comes up
showing missing classes.

This happens because the IDE may try to restore editor windows
*before* the Gradle machinery has set the AndroidModel field
in AndroidFacet. The net result is that the ModuleClassLoader
fails to add all the library jars needed for its dependencies,
and for layouts and themes that reference these views, the
rendering ends up broken.

The solution is to postpone module renders in that case; we
keep enqueing render requests until the model is present.

Change-Id: I9f9da0aa92c8fa419feeee39d5caf72709169ce9