Add live preview for canvas resizing
This only works for the preview tab (not the design one). In the design
tab, it will only display the resize overlay as it was doing before. The
reason for that is that, because resizing the canvas will automatically
change tabs to the best possible file, it breaks the interaction and the
live preview. It will also reset zoom levels and reposition the screens.

Change-Id: If8ba2aa990a1f70db1f19e7fbaf13046a977a949