Remove the designer reference from the palette panel.
Currently the PalettePanel keeps a reference to the designer from
when the palette is created. This will become stale after switching
to a different file. Remove the reference and add methods to the
design surface such that we get the functionality we want.
This also makes the palette switch to the correct components
when changing to a different type of xml file e.g. menu to
preference or layout.

Change-Id: I1770c246900a3280fddb3c6c0b5cce78b070067d
(cherry picked from commit 228654e545a70ce16c5112a888afb95738eea8f7)