209773: Quickfix for missing resource does not work
Invoking the quickfix for a missing resource value
( android:text="@string/unknown" )
does not work - regardless of the value you specify it will
just say that the resource name hasn't been defined.

The problem seems to be that it doesn't handle the case
where chooseName=false is passed into the dialog
constructor: if chooseName is false, it must remember
the name, and the place it expects to read it from is
the text field.

Change-Id: I5f4d60b6c53394610bb1118e1d95ec75928a811c
(cherry picked from commit 3a3257dbcc0ede9a739dd17ac1de120041884438)