[New PSD] Fixed scrolling issue in "Details" view, once and for all.
This CL finally fixes the scrolling issue in the "Details" view.
Previous attempts partially fixed the issue where this view was
scrolled down to the bottom if there were too many messages to show.

After reading Swing code, what the original fix was missing was
an "invokeLater" for setting the value of the vertical scroll bar,
so it is applied after the document in the JEditorPane does its
own scrolling when setting the caret position equal to zero.

Change-Id: I8e0f29bfe96daace36b4a528f3d68646edef08fd