Turn off auto-zoom on canvas resize
When you make the canvas smaller, the auto-fit right after
the resize would make the canvas bigger again (but zoomed in
make the pixels larger). I thought this was a bug at first until
I realized what was happening.

When you're resizing the canvas it's probably better to leave
the zoom at the same factor so the resize matches your mental
model of the new size relative to the previous one; there's
a button (and shortcut, now shown in the action tooltip) to
fit manually.

Change-Id: I173252e4d83632c5e101b52248143edbd51b2760