209963: System preference rendering doesn't work with appcompat
Add workaround such that SwitchPreferenceCompat views show
up in preference screens.

We need to do a bit more to support proper *editing* for
these (I tried making the ViewHandlerManager return the
non-compat handler, but that's not enough since (a) the property
sheet maps properties from the underlying XmlTag, (b) the outline
bases the icon on the non-compat name, and (c) the drop handler
should automatically use compat or non-compat versions based
on the library dependencies.

But with this CL you can at least see the system preferences.

Change-Id: I9d1a518a434448c4d6270b9bc3193b70ee923371