Fix compressed size computation in the presence of zip files
A recent CL () allowed viewing the contents of a zip file inside an APK
file. However, this caused a regression in the computation of compressed
file sizes since that routine assumed that every entry in the tree will
have a matching file inside the apk, but that is incorrect for entries
that correspond to files within the zip.

This CL utilizes a simple fix in such a scenario: the compressed file
sizes are marked as unknown for such files (those inside a zip inside
the apk).

Change-Id: I673db9289e62e3a9f2beaf613715690eb1d36b8b
(cherry picked from commit 1f701ab4eedc14fc5c6bb71dae70a941dba3ed7c)