[New PSD] Fixed "Resolved Dependencies" scrolling on selection.
With this CL, when a dependency is selected in the "Declared
Dependencies" panel and a matching dependency is selected
on the "Resolved Dependencies" panel, the tree will scroll to
the first selected element. Before this CL, there was no
scrolling and it was confusing to see what was selected in
the "Resolved Dependencies" panel, especially when the tree
view in that panel has a lot of elements.

Change-Id: Iba221db3de4f5b51801335680f09e520056ccac1
(cherry picked from commit 8590b1d4fc31550eff52d0fd135f3f0beb704cd9)