Fixing NPE when install location is changed
When the SDK install location is changed in the welcome wizard, we
create a new RepoManager to pick up the new path. This clears out the
cached packages, including the remotes. Then when we go to try to
resolve dependencies, we can't look up the remotes, and NPE.
Reloading the remotes before trying to resolve (using the default
timeout, so it won't happen more than needed).

Fixes b.android.com/210150

Change-Id: I2543502028a509d9c03521dc41a62f316b7dea58