Disable new dependency graph resolution.
This CL disables the code that tells the Gradle plugin
to return a full dependency graph. This functionality was
introduced in version 2.2.0 and apparently transitive
dependencies in Java library modules are not being
resolved properly.

This CL forces the plugin to return dependencies in
a list, just like in previous versions, until we figure
out the issue with the dependency graph.

Change-Id: Ic510d25f4d551ea36ecc822b360fdb816f1cfe4b