Adding SeriesDataList, and SeriesDataStore interfaces.
This change allows us to pull test data, and real data from outsite the
UI framework and push that data to the UI. In this change the simple
and most common usages of the continous series have been adapted. A
follow up change will migrage over the event, and state chart data.

Change-Id: Ie5ca7e1c0f1da02953b71ca02799c43f6c145de7