device picker: fix memory leaks
This CL fixes a few memory leaks:
 - The device picker provides an API for the higher level container to
   install a double click listener. But it holds on to that reference
   and never disposes it leading to the container staying in memory.
 - A few other listeners (e.g hyper link listener) were anonymous
   classes that were holding a reference to the enclosing class. They
   are now either lambdas, or static inner classes.
 - The loading panel used the project as a disposable, when it should
   have been using the disposable appropriate to the dialog lifetime.

It turns out that the dialog itself may still remain in memory since it
is held by a WeakReference inside Swing's JFrame class. So it is
expected that when looking at all the Windows inside a Frame, the test
infrastructure should look for the ones that are explicitly visible.

Ref: https://code.google.com/p/android/issues/detail?id=211383

Change-Id: I04a50efa701f2eed2f61fa1abf6903f22b87a0ec