Provide a dropdown for the style property.
Limit the styles to the styles that derive from the tag dedicated
styles i.e. Widget.Button for buttons etc.
The possible base style name is supplied by the ViewHandler.

Also in this CL: allow the NlEnumEditor to have a browse button.
We will display a browse button for styles since there may be a
valid style somewhere that does not derive from a known dedecated
style.

Change-Id: Ic0da5a482681fa63b5543d9f186b165da47ffecb