Add ThreadSelectedListener to ThreadsSegment
Add a listener to changes in the threads selected in the list. This
listener is useful for displaying information in other segments that are
based on the threads selected in ThreadsSegment. For instance, if one or
more threads are selected, another segment, containing flame charts,
bottom up calls and top down calls corresponding to the selected threads
should be displayed.

Change-Id: I8ab7ba270cd71d6a668b71e31d180f497d1583d9