Cleanup retrieving auth token from build-info.xml
A previous CL (10eb91238e3c) introduced a different mechanism for
obtaining the authorization key: rather than the IDE injecting a key,
Gradle will generate a key and persist it in the build-info.xml.

As a result, there is no need for the IDE to ever generate a secret key
- it should always read it from build-info.xml. This CL cleans up
existing code around the handling of the secret key.

Change-Id: I6122de4e9403d39fc8ab0b0b996159f10e1e69de