Temporarily remove support for auto restart build on run-as failures
When we fail to push dex files due to run-as failures, we attempt to
restart the build. This is implemented in a sub optimal way (See
https://b.android.com/212034 for details).

This CL temporarily removes the existing support as this also
complicates integration for external build systems. A different scheme
will be introduced in a future CL (tracked by the Issue 212034)

Change-Id: Id61ad24cc385bafe562ec5985f6aa87371de72a4