Cleanup IR deploy phase
This CL changes the deploy phase for instant run. The primary objective
is to decouple the project setup (gradle module) from the build result.
A new class InstantRunBuildResult is now responsible for analyzing the
output of an instant run build, and providing a list of tasks that need
to be invoked to update the application on the device and the local
instant run state.

The new architecture also simplifies how notifications are displayed -
although all the existing notifications are not plugged in as yet.

Change-Id: Iabc7c79d550ae369a26418b9db71d3ee19f39130