maybe fix vfs already exists error
During testing, I kept getting errors claiming that some
`.gfxtrace.temp` file already existed in the VFS.

As far as I can tell, we check to make sure that the trace
file itself doesn't exist, but we don't check the temp file.
I reckon the error happens when we've got leftover temp files
from a failed run, and we've deleted the trace file but not the temps.

Follow up question: should we use something like `File.deleteOnExit()`
with this?

Change-Id: I5716052a06f264b511b2cb4265dbae3c60bd4831