Only show non-palette view classes in <view> class drop handler
This CL adds support for clients of the OpenClassDialog to supply a
custom filter where they can filter out classes that shouldn't be
included in the list presented to the user.

This is used from the "view" tag handler to only show views that
aren't in the android namespace, since these are typically *not*
what you'd set using a <view> tag (besides they're mostly available
in the palette directly already). This makes the dialog quite a
bit less noisy.

Also change the palette name of the <view> tag to "<view>" (to
match the items it's next to -- <requestFocus> and <merge>, and
to indicate that this is a special layout inflater tag, not a
regular class (and "View" in particular was problematic since that
matches an actual view.)

Change-Id: I7f60f4279745f12a01f11009d239c6fe8a11b130
(cherry picked from commit 2c93bede504b58aa0616e6b8d235068bf58d750c)