Not downloading the so-called "sdk seed" in the first run wizard
Prior to this change, the first run wizard first downloaded a zip containing
the tools sdk package version 22.6.2 (current is 25.x) and some other random
unimportant stuff and installed that, and then went on to download the most
recent tools version separately.
This change removes the initial download of the old zip.

Fixes b.android.com/201235

Change-Id: I76021a931ca0f192601f01362583e33615c8711d