Added the possibility to display a Mockup
Added the possibility to add a mockup on the top component of a layout
using the "tools:mockup" attribute.
The mockup is displayed as a layer on top of the blueprint ScreenView.
Its position and cropping can be set through the "tools:mockup_position"
attribute with a string having the following form :
	x, y, [w, h, [cx, cy, cw, ch]]
 * x : x offset of the mockup on the ScreenView (in dip)
 * y : y offset of the mockup on the ScreenView (in dip)
 * w : width of the mockup on the ScreenView (in dip). will be scaled if needed
 * h : height of the mockup on the ScreenView (in dip). will be scaled if needed
 * cx : x offset of cropping area on the mockup (in px)
 * cy : y offset of cropping area on the mockup (in px)
 * cw : width of cropping area on the mockup  (in px)
 * cy : height of cropping area on the mockup  (in px)

Added a dialog edit the opacity and file, opened with a new button on
the toolbar. Will add the resizing later.

Change-Id: I1467bb1473a163349908637ccd1ad15173f0b190