Fix for Atom List having a size of 0 after switching Tabs.
The problem is that the main EditorTabs of the IDE likes to set
its children to 0 when it is not using them,

RunnerContentUi.initUi sets a border of size -1 to the "wrappper" component
This causes the layout to set its children to size of 1, as the border of
-1 makes it think it has space of 1, even when its total size is 0.

Splitter.doLayout would set the size of all its children to 0 if it size
is more then 0, and because of the -1 border, the size of the child is 1.

Then GridImpl.addNotify happens, and size is 0 so it moves the dividers
to the 0 position. making it look like the Atom list is gone.

This fix gets rid of the -1 boder, so the Splitter.doLayout does not set
its children to 0, so the GridImpl.addNotify does not mess up the divider
position.

You may be thinking, how come this problem does not happen the very first
time when we open the Trace View, but actually it does, but there is
already a hack inside RunnerContentUi.MyComponent.addNotify to fix it.
but there is a boolean myUiLastStateWasRestored, that means this hack
only runs the very first time the UI is created, as at that point it
most certainly will have a size of 0, and after it should not.

Change-Id: Id79dbf643c04023c4aeb2a4b8bd51c9068c541a1