Temporarily ignore test until failure identified.
This test was broke by a CL that we have had trouble identifying the
reason for. There was a project added, and we had a huge influx of CL's
that hid on the bots what caused this test to fail. We know it started
to fail around 06/09 around 2PM PDT. The test seems to have become
flaky.

Change-Id: Iff646949d4413c2ba68efa53063c6860005770ba