Terminate launch if hot or cold swap update fails
With the current API's, certain errors that could happen while the IDE
was pushing patches only resulted in notifications (which could be
turned off), but did not terminate the launch itself. This CL changes
the launch flow to terminate a launch when there are errors while
pushing patches to the device.

Fixes https://code.google.com/p/android/issues/detail?id=211989

Change-Id: Ib8b417fe56a1f03b55c0228d45523e65436172eb