Fix instant run notifications
Instant Run notifications used to come from a number of different
locations like the run configuration, the before run task, the
callbacks during various artifact pushes, and finally at the end of the
launch tasks.

This CL simplifies (in conjunction with a previous CLs and changes in
Gradle) all of that to two locations - the run configuration when it
decides that the current launch should not use instant run, and a new
notification task that is performed after all the instant run specific
launch tasks.

Fixes https://code.google.com/p/android/issues/detail?id=211506 and
https://code.google.com/p/android/issues/detail?id=205845

Change-Id: I2ae5d669a5b24756fff9d7298a9446c8ef3bf05b