Re-introduce support for detecting broken 'run-as' for IR
Instant Run relies on having a working run-as command when doing a
freeze swap. If run-as is broken, we need to do a full build instead.
A previous CL (826ea399) removed the existing support for detecting this
since it was quite convoluted: we relied on installing the patches,
caught the error on failure, then restarted a build.

This CL re-introduces the support in a much simpler way: We know that we
only need this for the freeze swap scenario. This implies that we can
rely on the application already being installed on the device. As a
result, we can check whether a device has a working run-as by examining
the output of a simple command such as: "run-as com.foo ls"

Fixes https://code.google.com/p/android/issues/detail?id=212034 and
https://code.google.com/p/android/issues/detail?id=207347

Change-Id: I0ff526adf49dcae3051a7cb5a32c72de8339d358