Fix layout editor zoom behavior under retina mode
In Retina mode, we double the zoom scale in order to paint Retina
images. However, that caused the zoom in/zoom out behavior to be
weird, since it was performing the jumps on the doubled percentages
instead of the hardcoded list of 90, 100, 110, 125, 150, etc.

Also, the special case to avoid scaling when we're at exactly 100%
was not correct for Retina.

Change-Id: Iaa304d9657b011d47b4f25db83eee62a97e4a17f
(cherry picked from commit 2df789113167fc6bd06883d5a64818dca84cff64)