Adding DataAdapter to SeriesDataStoreImpl.
Adding the concept of a DataAdapter to the DataStore, each SeriesDataType
that we collect should have a DataAdapter that we can then use to store,
and poll data for use in the UI.
This change also combines 2 APIs that were always called in series and
replaced it with getDataAt that now returns the SeriesData. This ensures
all DataAdapters keep the data and timestamps insync.

Change-Id: I035c5a29794b547bb376ad9b853674b524f69091