Update the skin correctly after changing device.
Bug: http://b.android.com/209212
It looks like a refactoring of the AVD dialogs caused the skin to
be incorrectly determined. The problem seem to be that the system
image was unavailable at the point where the skin file was
resolved.

Change-Id: I2baa7b532ea85439ab0c981985f848bf07ae8293
(cherry picked from commit 45a5edc3feaf1fb3bcd730bfccb37f2a55f34ae7)