Fixing NullpointerException in EventComponents
This change reworks how event components draw their lines. Instead of
matching a start and end event, each component now uses the information from the
end event to determine how / where to draw itself. If a starting event does not
have an end event it will draw to the end.

Change-Id: I8fed8eeb1e5c00cc903a58a3be7b4a34df04650b