Query the running app for the ABI.
In the case of click-to-trace we were previously just using the
device's primary ABI. This would cause problems if the APK only
supported armv7, but the device's primary ABI was armv8.

Increased the JDWP timeout from 20 seconds to 60 as we're now
pushing the APKs inside this timeout window instead of before.

Fixes: 29457279

Change-Id: Iecb4f127ae2f389b089570200b6976c445b02ffd