Reshuffle GapiiLibraryLoader logic to fix race conditions.
We were accessing the DebugProcess’s RequestManager on a different
thread to the debugger, which caused raciness. We now use it on the
processAttached callback, which is just before the editor sets its
regular breakpoints.

Moved all usages of the debugger and the debugger session to the same
executor, again to avoid raciness.

Add more comments.

Fixes: 29479918
Change-Id: I7ecef54642ce58b77b5807898ef0d58ae5654aae