Exclude firebase crash receiver process from forcing cold swap
Instant Run forces cold swap if a process uses multiple processes.
However, we explicitly exclude scenarios where the second process
happens to have come in via a dependency, e.g. the app may depend on
leakcanary, which introduced the 2nd process.

This CL adds the firebase crash receiver service to the list of known
process names so that a cold swap is not forced simply because the user
added a dependency on play services. The easiest way to accidentally
trigger this dependency seems to be to create a wear module, which
forces a dependency on play-services, which seems to pull in a whole
bunch of dependencies include firebase-crash.

Fix for https://code.google.com/p/android/issues/detail?id=213494

Change-Id: I70aef0cdf3219c4aec251a87754e6e9da96a643e