instant run: Only allow reusing process handler for hotswaps
The instant run builder decides to force a cold swap if the app is not
in foreground. However, if there were no changes, then the analyzer
decides that the same process handler can be reused. The net effect is
that the app still stays in the background.

This CL fixes the issue by making sure that the analyzer takes the
builder's choice into account when deciding whether the process handler
can be reused. Specifically, the process handler can be reused only if
the builder performed an incremental hotswap build.

Also don't log warnings if connection to app fails when the app is not
running.

Fixes https://code.google.com/p/android/issues/detail?id=205011

Change-Id: Ib9ffe37e7b5cf1d4c7991d2ffc7b5d294a8b5b1f
(cherry picked from commit 52e50206d6d9229b42bde4b6303754ccab688be6)