Support multi-module services.
Adds support for mutliple service instances matching the same
service id due to module statefulness.
Also adds support for messages being present at the same time
as buttons in the case of partially complete actions.
Includes a few related refactors.

NOTE: There were some compromises in design and the expectation
is that this will be further refactored moving forward.

Risk: Medium, the code is isolated to the Firebase plugin and the Assistant (of which Firebase is the only consumer at this time). OTOH, this is a medium size refactor.
Justification: We do not handle multi-module or multi-project usage without this change. Dependencies may be applied to the incorrect module and button statefulness be inaccurate.

Bug: 28046271
Change-Id: I67d841f3227e89c3bc51707bffc0edbd635abc92
(cherry picked from commit 743b74453fb51cf9ee8c402653928cfa3041bc46)