IR: Check if APK is still installed for default user
When checking whether an APK is installed, we use "dumpsys package
<pkgName>" and examine its output. On multi user/restricted profile
systems, an app could be uninstalled from one user, and the dumpsys
package will indicate for which users it is still installed.

Instant Run however doesn't actually understand the concept of per user
installation right now. This CL makes sure that we check whether it is
still installed for the default user.

Eventually, we should either save the state about the user id on which
deployment happens, and make sure that the system considers it installed
for that particular user id.

Fixes https://code.google.com/p/android/issues/detail?id=199647

Change-Id: I430ecb2471fc3dc95d86ccd3c592bfef6f06c5e2