Link CpuDataPoller with SeriesDataStore
Store CPU data pulled from the device into the data store. As a result,
there's an initial end-to-end flow: data pulled from the device is being
displayed in studio.

However, there are some things that need to be done. There are some
drops that can be observed in the charts every second or so. A simple
app that triggers some processing via a button can be used for testing
the workflow.

Change-Id: I330821f3952568974d6211ba7f75df05464fa960