Fix the design button on properties with no browse button.
Prior to this change, the BrowsePanel was hidden when we detected
that a property should not have a resource picker.
That would hide the design button while editing this disabling the
"create design property" feature for properties like "alpha".

This is now fixed by letting the BrowsePanel decide what should
happen on a property change.

Change-Id: I37724d16b2ea3679a12450b550d4d15d86b9ee64