Call removeRunContent() in a ProcessListener.processTerminated()
As far as I can tell, the piece of code that was initially throwing
the exception mentioned in the bug, is executed on a listener registered
before we get a chance to register this one, so we shouldn't be able to
trigger it anymore. Running through the trace flow a few times, I
haven't managed to repro. Shrug.

Bug: 29508203
Change-Id: Ib4556fba524a32997de2d510046a5baf674611b5