Use MergedManifest when computing manifest resource hash
The logic to compute the hash of all app resources referenced from the
manifest used the first merged manifest it could find. This CL replaces
that to use the MergedManifest API instead. This allows us to not rely
on the merged manifest from the previous build.

Also includes a few minor refactorings and tests.

Change-Id: Iace1bb747fbca318c7091bf81de24925742e00e3