Fix NPE when loading an hprof file.
According to the trace, it seems like the field mySnapshot
was never assigned to. The only window this could happen
is when the Snapshot is being processed and the editor is
suddenly closed.

This change moves the computation into an intermediate
variable, and is only assigned to the member variable in
the EDT only when all computations have successfully
completed. This simplifies the error detection logic and
removes the need of synchronization primitives.

Change-Id: If1a65168a35bc5785160ad9ea0671c076dc2fd32