Reduce scope of read lock in ResourceFolderRepository#init
Attempt to reduce the chance of locking up the
EDT thread (write contending with read lock).

Remove read lock from main init scan since most
of this is done with VirtualFiles now. Queue
files that will be retried as PSI (data binding,
files with duplicate items, etc.), and handle
those in a second pass instead where we do
finally grab a read lock.

Stop grabbing stub PsiFiles for every file that
wasn't cached in the merged state. Previously
stub PsiFiles were loaded to ensure that the
PsiListener would get events. It turns out that
only the parent directories need to be PSI.
I'm not sure this is documented anywhere, but
attempt to test for it.

Misc cleanup to rename field.

BUG: b/29554795
Change-Id: Ie9ce2d17361ff75a05e363738307280058dc7c76