IR: Hash the contents of files for file based resources
When computing the hash of all resources referenced from the manifest,
we were incorrectly hashing the path to the resource rather than the
resource itself. This CL changes the hasher to hash the contents of such
file based resources.

This is one step better than what we used to have, but it is not fully
accurate either. In particular, file based resources may further
reference other resources, and hence a full transitive closure of all
resources need to be hashed. This is a stop gap measure until we fully
figure that out.

Partial fix for http://b.android.com/205075

Change-Id: Ia1327b8d109e77a11feede19818fd03d37910648