Manifest viewer: switch to file
Update the file click handler (used when you want to
for example look at the manifest file from play services)
such that it attempts to switch to the text editor, if
available. Without this, you'd end up just looking at
the merged view for that file, which isn't very helpful.

Change-Id: Ibf2865c5478db40b98376407cbc1b558c09096ea