Dependencies in Android projects are no longer "exported".
Before AS 2.2 we relied in dependencies to be "exported", so
things like jar files in "libs" directory were magically
available to dependent modules.
In AS 2.2, IDEA no longer *exports* "exported dependencies
in Gradle projects (see GradleOrderEnumeratorHandler. Thanks,
Fuyao, for finding the root of the problems!).
Because of this change in behavior, we explicitly set up all
types of dependency, without assuming that the IDE will
"export" them. This gives us more control and knowledge of
what is going on.
Since we no longer rely on "exported" dependencies, this CL
removes any invocations to "setExported(true)" when setting
up dependencies.

Change-Id: I1721120adce237a84b87144b44ee41ee350e938c